feat(core): Write temp dataset file as JSONL to avoid escaping issues (#123)
When creating a dataset we first save it to a temporary file. If the
content is JSON strings then they need to be escaped properly with
double quotes, which isn't a very good UX. This changes the logic to
save the data as a temporary JSONL file instead, which doesn't require
any special escaping logic.